sygnalizacja swietlna
nowe assety i nowy skrypt
teraz jest sygnalizacja swietlna
aby ja dodac trzeba umiescic prefab swiatel w obiekcie obst  to wazne

+++ Assets/script/auta/obstacleControler.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class obstacleControler : MonoBehaviour
 {
     public bool isActive;
-    public int waitFor;
+    public float waitFor;
     private void Start()
     {
         if (waitFor == 0)
